refactor: remove musig status checks (#1886)
