+++ .github/workflows/test.yml
@@ -843,10 +843,10 @@ jobs:
             - name: Test
               uses: nick-invision/retry@v2
               env:
-                  COVERAGE_THRESHOLD_LINES: 99.93
-                  COVERAGE_THRESHOLD_FUNCTIONS: 99.75
-                  COVERAGE_THRESHOLD_STATEMENTS: 99.38
-                  COVERAGE_THRESHOLD_BRANCHES: 99.38
+                  COVERAGE_THRESHOLD_LINES: 98.76
+                  COVERAGE_THRESHOLD_FUNCTIONS: 99.53
+                  COVERAGE_THRESHOLD_STATEMENTS: 98.8
+                  COVERAGE_THRESHOLD_BRANCHES: 97.17
                   COVERAGE_INCLUDE_PATH: src/domains/wallet
               with:
                   timeout_minutes: 10
