+++ src/domains/wallet/pages/WalletDetails/hooks/use-wallet-options.test.tsx
@@ -82,7 +82,6 @@ describe("Wallet Options Hook", () => {
 			options: [
 				{ label: "Delegate", value: "delegate-registration" },
 				{ label: "Second Signature", value: "second-signature" },
-				{ label: "Multisignature", value: "multi-signature" },
 			],
 			title: "Register",
 		});
@@ -96,10 +95,7 @@ describe("Wallet Options Hook", () => {
 
 		expect(result.current.registrationOptions).toStrictEqual({
 			key: "registrations",
-			options: [
-				{ label: "Delegate", value: "delegate-registration" },
-				{ label: "Multisignature", value: "multi-signature" },
-			],
+			options: [{ label: "Delegate", value: "delegate-registration" }],
 			title: "Register",
 		});
